avoid some unnecessary rebuilds
Running autoreconf on the vendor/ dir causes its timestamp to change,
which tells cargo to re-run the build script every time.

Ideally, we would have an intermediate source preparation step where we
run autoreconf only once. We don't need to run this as part of the build
step.

Another solution would be to include directly a distribution tarball in
the git repo, but that comes with its own set of cons.

So for now I'm settling on a quick solution, i.e. avoiding the `-f`
parameter. This way autoreconf only modifies the vendor/ directory once.

(I would still love to eventually find a way to skip the autoreconf
completely, as this wastes several seconds during each build)