fix broken ios build
The integration with autotools was failing to detect when the package is
being built for iOS, with the result that it was linked against the
wrong sdk.

At this point the simplest solution is to compile the few cpp files
directly using the `cc` crate, which is indeed able to handle all the
edge cases of iOS and Android builds.